requirements.txt
boto3==1.26.142
httpx==0.24.1
nest-asyncio==1.5.6
pydantic==1.10.9
PyPDF2==3.0.1
llama-index==0.7.7
langchain==0.0.217
openai==0.27.8
matplotlib==3.7.1
numpy==1.24.3
scikit-learn==1.2.2
python-dotenv==1.0.0
office365-rest-python-client==2.4.1
atlassian-python-api==3.36.0